tagi

+++ app/src/Controller/CategoryController.php
@@ -189,6 +189,12 @@ class CategoryController extends AbstractController
      */
     public function delete(Request $request, Category $category, CategoryRepository $categoryRepository): Response
     {
+        if($category->getQuestions()->count()){
+            $this->addFlash('warning', 'message_category_contains_questions');
+
+            return $this->redirectToRoute('category_index');
+        }
+
         $form = $this->createForm(FormType::class, $category, ['method' => 'DELETE']);
         $form->handleRequest($request);
